backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
streamlit==1.39.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
redis==5.2.0
celery==5.4.0
praw==7.7.1
telethon==1.36.0
feedparser==6.0.10
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
trafilatura==1.10.0
dateparser==1.2.0
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0